src/requirements.txt
auto_mix_prep==0.2.0
joblib==1.5.3
pandas==2.3.3
PyYAML==6.0.3
Requests==2.32.5
scikit_learn==1.8.0
tabulate==0.9.0
